完成购物车界面，完全实现美团外卖的购物车界面的所有逻辑，接下来开始实现订单界面，此项目的预期主体功能完成已经达到一半多了

+++ app/src/main/java/com/example/axiang/warmstomach/adapters/HomeAdapter.java
@@ -281,7 +281,7 @@ public class HomeAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
         viewHolder.itemStoreName.setText(store.getStoreName());
         viewHolder.itemStoreType.setText(store.getStoreType());
         viewHolder.itemStoreMonthSoldOn.setText(store.getStoreMonthSoldOn().intValue() + "");
-        viewHolder.itemStoreStartingPrice.setText(store.getStoreStartingPrice().intValue() + "");
+        viewHolder.itemStoreStartingPrice.setText(store.getStoreStartingPrice().doubleValue() + "");
         viewHolder.itemStoreDelivery.setText(store.getStoreDeliveryFee().floatValue() + "");
         viewHolder.itemStorePerCapita.setText(store.getStorePerCapita().intValue() + "");
         if (store.getStoreOnlinePayment() == null || !store.getStoreOnlinePayment().booleanValue()) {
